requirements.txt
asgiref==3.8.1
Django==5.1.6
django-cors-headers==4.7.0
django-jazzmin==3.0.1
django-multiselectfield==0.1.13
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
sqlparse==0.5.3
tzdata==2025.1
channels==4.0.0
daphne==4.1.0
channels-redis==4.1.0
